+++ src/game/frontend/items/Items.hpp
@@ -108,7 +108,7 @@ namespace YimMenu
 	private:
 		ListCommand* m_Command;
 		std::optional<std::string> m_LabelOverride;
-		std::optional<int> m_ItemWidth = std::nullopt;
+		std::optional<float> m_ItemWidth = std::nullopt;
 		std::optional<std::string> m_SelectedItem = std::nullopt;
 	};
